requirements.txt
fastapi==0.103.1
pydantic==2.2.1
pandas==2.0.3
numpy==1.24.3
openai==1.56.1
rapidfuzz==3.1.1
xgboost==2.1.3
lightgbm==4.0.0
joblib==1.3.1
scikit-learn==1.2.2
matplotlib==3.7.2
shap==0.42.0
python-dotenv==1.0.1
uvicorn==0.32.1
pyarrow==18.1.0